% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_calc_markers.R
\name{calculate_cluster_markers}
\alias{calculate_cluster_markers}
\title{Calculate Cluster Markers for a COTAN Object}
\usage{
calculate_cluster_markers(
  cotan_obj,
  output_dir = NULL,
  file_name = "cotan_markers.rds"
)
}
\arguments{
\item{cotan_obj}{A \code{COTAN} object with uniform clustering already applied.}

\item{output_dir}{Character. Optional directory to save the output object as an RDS file. Default is \code{NULL}.}

\item{file_name}{Character. Optional file name for saving the output object. Default is \code{"cotan_markers.rds"}.}
}
\value{
A list containing the updated \code{COTAN} object and the cluster markers data.frame.
}
\description{
Performs differential expression analysis on clusters and extracts the markers.
}
