% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_conv_sce.R
\name{convert_to_sce}
\alias{convert_to_sce}
\title{Convert COTAN Object to SingleCellExperiment}
\usage{
convert_to_sce(cotan_obj)
}
\arguments{
\item{cotan_obj}{A \code{COTAN} object.}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Wraps the COTAN object into a SingleCellExperiment object for downstream interoperability.
}
